Menu em loop
troquei as condicionais do menu por switch case e implementei o loop do menu e a opção sair

+++ bolos.h
@@ -9,11 +9,11 @@ struct data{
 
 struct cliente{
   char nome[max_length_nome];
-  long int cpf;
+  unsigned long long int cpf;
 };
 
 struct pedido{
   struct data quando;
-  long int cpf;
+  unsigned long long int cpf;
   char tipo_bolo[max_length_bolo];
 };
